% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayescript_profiles.R
\name{bayescript_profiles}
\alias{bayescript_profiles}
\title{Builds script to run all analyses from ontobayes_profiles.}
\usage{
bayescript_profiles(tags = tags, OS = "windows", filename = "bayescript")
}
\arguments{
\item{tags}{list, profiles with anatomy ontology terms to query. The list of profiles should be organized as indicated in the examples.}

\item{OS}{character, sets the operational system to produce batch files for. Options are "windows" and "unix". Default is set to "windows".}

\item{filename}{character, name of the output file.}
}
\value{
A script to run all MrBayes batch files inside NEXUS subfolders.
}
\description{
Builds script to run all analyses from ontobayes_profiles.
}
\examples{
\dontrun{
# HAO example #
# List of profiles with terms to query #
tags <- list(
  c("HAO:0000506", "HAO:0000513", "HAO:0000453"),
  c("HAO:0000234", "HAO:0001003"),
  c("HAO:0000874", "HAO:0000583", "HAO:0000630"))

# Set profile names and terms #
names(tags) <- c("mouthparts_profile1", "head_profile2", "mesosoma_profile3")
names(tags[[1]]) <- c("mandible", "maxilla", "labium")
names(tags[[2]]) <- c("cranium", "tentorium")
names(tags[[3]]) <- c("prothorax", "mesothorax", "metathorax")

# Create MrBayes batch files to run all analyses #
bayescript_profiles(tags = tags, OS = "windows", filename = "bayesprof1")

# UBERON example #
# List of profiles with terms to query #
tags <- list(
  c("UBERON_0002244", "UBERON_0002397", "UBERON_0004742"),
  c("UBERON_2000658", "UBERON_2000488"),
  c("UBERON_0000151", "UBERON_0000152", "UBERON_0003097"))

# Set profile names and terms #
names(tags) <- c("jaw_profile1", "pharynx_profile2", "fins_profile3")
names(tags[[1]]) <- c("premaxilla", "maxilla", "dentary")
names(tags[[2]]) <- c("epibranchial_bone", "ceratobranchial_bone")
names(tags[[3]]) <- c("pectoral_fin", "pelvic_fin", "dorsal_fin")

# Create MrBayes batch files to run all analyses #
bayescript_profiles(tags = tags, OS = "windows", filename = "bayesprof2")
}

}
