% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesinfo_profiles.R
\name{bayesinfo_profiles}
\alias{bayesinfo_profiles}
\title{Import and summarize results from analyses set by the galaxscrip_profiles function.}
\usage{
bayesinfo_profiles(tags = tags, runs = 2)
}
\arguments{
\item{tags}{list, profiles with anatomy ontology terms to query. The list of profiles should be organized as indicated in the examples.}

\item{runs}{integer, the number of independent runs of mcmc used in the MrBayes commands block. Default is set to 2.}
}
\value{
A list with matrices. Each matrix shows values of posterior coverage, Bayesian phylogenetic information, and phylogenetic dissonance
  among mcmc runs for each subset of data defined by an anatomy ontology term and across all subsets for each profile defined.
}
\description{
Import and summarize results from analyses set by the galaxscrip_profiles function.
}
\examples{
\dontrun{
# HAO example #
# List of profiles with terms to query #
tags <- list(
  c("HAO:0000506", "HAO:0000513", "HAO:0000453"),
  c("HAO:0000234", "HAO:0001003"),
  c("HAO:0000874", "HAO:0000583", "HAO:0000630"))

# Set profile names and terms #
names(tags) <- c("mouthparts_profile1", "head_profile2", "mesosoma_profile3")
names(tags[[1]]) <- c("mandible", "maxilla", "labium")
names(tags[[2]]) <- c("cranium", "tentorium")
names(tags[[3]]) <- c("prothorax", "mesothorax", "metathorax")

# Import and summarize results from profiles #
profs1 <- bayesinfo_profiles(tags = tags, runs = 2)

# UBERON example #
# List of profiles with terms to query #
tags <- list(
  c("UBERON_0002244", "UBERON_0002397", "UBERON_0004742"),
  c("UBERON_2000658", "UBERON_2000488"),
  c("UBERON_0000151", "UBERON_0000152", "UBERON_0003097"))

# Set profile names and terms #
names(tags) <- c("jaw_profile1", "pharynx_profile2", "fins_profile3")
names(tags[[1]]) <- c("premaxilla", "maxilla", "dentary")
names(tags[[2]]) <- c("epibranchial_bone", "ceratobranchial_bone")
names(tags[[3]]) <- c("pectoral_fin", "pelvic_fin", "dorsal_fin")

# Import and summarize results from profiles #
profs2 <- bayesinfo_profiles(tags = tags, runs = 2)
}

}
