% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxscript_profiles.R
\name{galaxscript_profiles}
\alias{galaxscript_profiles}
\title{Builds all input files and an script to run all analyses from ontobayes_profiles in Galax.}
\usage{
galaxscript_profiles(
  tags = tags,
  runs = 2,
  burnin = 251,
  OS = "windows",
  filename = "galaxscript"
)
}
\arguments{
\item{tags}{list, profiles with anatomy ontology terms to query. The list of profiles should be organized as indicated in the examples.}

\item{runs}{integer, the number of independent runs of mcmc used in the MrBayes commands block. Default is set to 2.}

\item{burnin}{integer, the absolute number of trees as the burn-in used in the mcmc. Default is set to 251.}

\item{OS}{character, sets the operational system to produce batch files for. Options are "windows" and "unix". Default is set to "windows".}

\item{filename}{character, name of the output file.}
}
\value{
All necessary input files, organized in subfolders inside the GALAX/INPUT folder, 
  and an script to run all Galax analyses.
}
\description{
Builds all input files and an script to run all analyses from ontobayes_profiles in Galax.
}
\examples{
\dontrun{
# HAO example #
# List of profiles with terms to query #
tags <- list(
  c("HAO:0000506", "HAO:0000513", "HAO:0000453"),
  c("HAO:0000234", "HAO:0001003"),
  c("HAO:0000874", "HAO:0000583", "HAO:0000630"))

# Set profile names and terms #
names(tags) <- c("mouthparts_profile1", "head_profile2", "mesosoma_profile3")
names(tags[[1]]) <- c("mandible", "maxilla", "labium")
names(tags[[2]]) <- c("cranium", "tentorium")
names(tags[[3]]) <- c("prothorax", "mesothorax", "metathorax")

# Creates bat file to run all analyses on Galax #
galaxscript_profiles(tags = tags, runs = 2, burnin = 251, OS = "windows", filename = "galaxprof1")

# UBERON example #
# List of profiles with terms to query #
tags <- list(
  c("UBERON_0002244", "UBERON_0002397", "UBERON_0004742"),
  c("UBERON_2000658", "UBERON_2000488"),
  c("UBERON_0000151", "UBERON_0000152", "UBERON_0003097"))

# Set profile names and terms #
names(tags) <- c("jaw_profile1", "pharynx_profile2", "fins_profile3")
names(tags[[1]]) <- c("premaxilla", "maxilla", "dentary")
names(tags[[2]]) <- c("epibranchial_bone", "ceratobranchial_bone")
names(tags[[3]]) <- c("pectoral_fin", "pelvic_fin", "dorsal_fin")

# Creates bat file to run all analyses on Galax #
galaxscript_profiles(tags = tags, runs = 2, burnin = 251, OS = "windows", filename = "galaxprof2")
}

}
