% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxscript_structure.R
\name{galaxscript_structure}
\alias{galaxscript_structure}
\title{Builds all input files and an script to run all analyses in Galax for comparisons between partitions according to the ontology structure 
  represented as a semantic similarity dendrogram or another type of reference dendrogram (e.g., based on phylogenetic dissonance).}
\usage{
galaxscript_structure(
  struc = struc,
  ids = ids,
  runs = 2,
  burnin = 251,
  OS = "windows",
  filename = "galaxscript"
)
}
\arguments{
\item{struc}{list, the output from the ontostructure function.}

\item{ids}{data.frame, matches between characters and anatomical entities. 
The first column should provide character ids in the same order as presented in the character matrix. 
The second column should provide the respective ontology ids referring to the anatomical entities. 
The third column should provide the respective ontology terms referring to the anatomical entities.
See examples provided in the data folder (data1.csv and data2.csv).}

\item{runs}{integer, the number of independent runs of mcmc used in the MrBayes commands block. Default is set to 2.}

\item{burnin}{integer, the absolute number of trees as the burn-in used in the mcmc. Default is set to 251.}

\item{OS}{character, sets the operational system to produce batch files for. Options are "windows" and "unix". Default is set to "windows".}

\item{filename}{character, name of the output file.}
}
\value{
All necessary input files, organized in subfolders inside the GALAX/INPUT/structure folder, 
  and an script to run all Galax analyses.
}
\description{
Builds all input files and an script to run all analyses in Galax for comparisons between partitions according to the ontology structure 
  represented as a semantic similarity dendrogram or another type of reference dendrogram (e.g., based on phylogenetic dissonance).
}
\examples{
\dontrun{
# HAO example #
# Create ID object #
ID <- as.data.frame(cbind(paste0("C", 1:40),
c(rep("HAO:0000506",5), rep("HAO:0000513",5), rep("HAO:0000453",5), rep("HAO:0000234",5), 
rep("HAO:0001003",5), rep("HAO:0000874",5), rep("HAO:0000583",5), rep("HAO:0000630",5)),
c(rep("mandible",5), rep("maxilla",5), rep("labium",5), rep("cranium",5)
, rep("tentorium",5), rep("prothorax",5), rep("mesothorax",5), rep("metathorax",5))))

# Import and visualize results from pairwise comparisons of terms #
pp <- bayesinfo_pairwise(ids = ID, runs = 2, profile = F, cluster = T)

# Get hierarchical structure based on dissonance dendrogram #
hstruc1 <- ontostructure(dsm = pp, manual = T, plot = F)

# Files to run analyses based on the dissonance dendrogram #
galaxscript_structure(struc = hstruc1, ids = ID, runs = 2, burnin = 251, OS = "windows", filename = "galaxstr1")

# UBERON example #
# Create ID object #
ID <- as.data.frame(cbind(paste0("C", 1:40),
c(rep("UBERON_0002244",5), rep("UBERON_0002397",5), rep("UBERON_0004742",5), rep("UBERON_2000658",5), 
rep("UBERON_2000488",5), rep("UBERON_0000151",5), rep("UBERON_0000152",5), rep("UBERON_0003097",5)),
c(rep("premaxilla",5), rep("maxilla",5), rep("dentary",5), rep("epibranchial bone",5)
, rep("ceratobranchial bone",5), rep("pectoral fin",5), rep("pelvic fin",5), rep("dorsal fin",5))))

# Get hierarchical structure based on semantic similarity dendrogram #
hstruc2 <- ontostructure(ids = ID, manual = F, plot = F, similarity = "jaccard")

# Files to run analyses based on the semantic similarity dendrogram #
galaxscript_structure(struc = hstruc2, ids = ID, runs = 2, burnin = 251, OS = "windows", filename = "galaxstr2")
}

}
