% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bayesinfoHM.R
\name{get_bayesinfoHM}
\alias{get_bayesinfoHM}
\title{Get node specific componenets of Bayesian phylogenetic information, posterior probability, and internode confidence for all majority rule trees 
  obtained from the analyses of data subsets defined by anatomy ontology terms.}
\usage{
get_bayesinfoHM(phy = phy, foldername = "all_partitions")
}
\arguments{
\item{foldername}{character, the name of the Galax output folder with the results from ontobayes analysis.}

\item{phylo, }{a reference phylogenetic species tree.}
}
\value{
A list of matrices with values of Bayesian phylogenetic information, posterior probability, and internode confidence  
  for each node in the reference phylogenetic species tree.
}
\description{
Get node specific componenets of Bayesian phylogenetic information, posterior probability, and internode confidence for all majority rule trees 
  obtained from the analyses of data subsets defined by anatomy ontology terms.
}
\examples{
\dontrun{
# Example: HAO or UBERON #
# Import a reference phylogenetic species tree #
reftree <- ape:::read.nexus(file = "./data/reftree.tre")

# Get node-specific components from each tree inferred from all different data subsets defined in ontobayes # 
M <- get_bayesinfoHM(phy = reftree, foldername = "all_partitions")
}

}
