% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp_bayes.R
\name{resamp_bayes}
\alias{resamp_bayes}
\title{Builds MrBayes NEXUS files for different types of resampling analyses.}
\usage{
resamp_bayes(
  ids = ids,
  type = "none",
  nsamp = 100,
  nchars = c(10, 10),
  ntax = 10,
  roots = roots,
  manual = T,
  ontology = "data/ontology.obo",
  relations = c("part_of", "is_a"),
  dataset = "data/matrix.nex",
  outgroup = "outgroup_taxon",
  rates = "gamma",
  coding = "all",
  ratepr = "fixed",
  brlenspr = "Unconstrained:Exp(10)",
  gen = 1e+06,
  runs = 2,
  chains = 4,
  freq = 1000,
  foldername = "resamp"
)
}
\arguments{
\item{ids}{data.frame, matches between characters and anatomical entities. The first column should provide character ids in the same order presented in the character matrix. 
The second column should provide the respective ontology ids referring to the anatomical entities. 
The third column should provide the respective ontology terms referring to the anatomical entities.
See examples provided in the data folder (data1.csv and data2.csv).}

\item{type}{character, sets the type of resampling analysis. Options are "none" to resample the same dataset defined in dataset; 
"char" to resample characters from the dataset defined in dataset; "tax" to resample taxa from the dataset defined in dataset;
"both" to resample characters and taxa from the dataset defined in dataset.}

\item{nsamp}{integer, sets the number of samples for the resampling analysis.}

\item{nchars}{integer, sets the number of characters to resample for each data subset if mode = "char" or "both". If more than one data subset
is selected then should be a vector with more than one element.}

\item{ntax}{integer, sets the number of taxa to resample if mode = "tax" or "both".}

\item{roots}{character, a named vector with the ontology IDs of the anatomical data subsets to produce MrBayes NEXUS files. 
IDs and names should match those in the second and third columns of the ids data.frame, respectively.}

\item{manual}{logical, if TRUE specify that an ontology file (OBO format) is provided by the user. This option is the only one currently available. Future developments will allow to access Uberon ontology directly using rphenoscape dependencies.}

\item{ontology}{character, name of the OBO file with the ontology.}

\item{relations}{character, set the types of ontological relations to use. Default is set to c("part_of", "is_a").}

\item{dataset}{character, name of the NEXUS file with the character matrix to be read.}

\item{outgroup}{character, name of the outgroup taxon.}

\item{rates}{character, sets the model of among-character rate variation in the MrBayes commands block. Default is set to "equal".}

\item{coding}{character, sets the coding bias in the MrBayes commands block. Default is set to "all".}

\item{ratepr}{character, sets the character specific or partition specific rates model in the MrBayes commands clock. Default is set to "fixed".}

\item{brlenspr}{character, sets the branch lengths prior in the MrBayes commands block. Deafault is set to "Unconstrained:Exp(10)".}

\item{gen}{integer, sets the number of generations of mcmc in the MrBayes commands block. Default is set to 1000000.}

\item{runs}{integer, sets the number of independent runs of mcmc in the MrBayes commands block. Default is set to 2.}

\item{chains}{integer, sets the number of chains for each run in the MrBayes commands block. Default is set to 4.}

\item{freq}{integer, sets the sampling frequency for the mcmc in the MrBayes commands block. Default is set to 1000.}

\item{foldername}{character name of the output folder to save all MrBayes NEXUS files.}
}
\value{
nsamp NEXUS files for each data subset specified in roots.
  The files will be organized in the foldername subfolder inside the folder NEXUS/RESAMP.
}
\description{
Builds MrBayes NEXUS files for different types of resampling analyses.
}
