% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp_galax.R
\name{resamp_galax}
\alias{resamp_galax}
\title{Builds all input files to run all analyses from resamp_bayes in Galax.}
\usage{
resamp_galax(
  roots = roots,
  nsamp = 100,
  runs = 4,
  burnin = 251,
  OS = "windows",
  foldername = "resamp",
  filename = "galaxscript"
)
}
\arguments{
\item{roots}{character, a named vector with the ontology IDs of the anatomical data subsets to produce MrBayes NEXUS files. 
IDs and names should match those in the second and third columns of the ids data.frame, respectively.}

\item{nsamp}{integer, sets the number of samples for the resampling analysis.}

\item{runs}{integer, sets the number of independent runs of mcmc in the MrBayes commands block. Default is set to 2.}

\item{burnin}{integer, sets the absolute number of trees as the burn-in for the mcmc. Default is set to 251.}

\item{OS}{character, sets the operational system to produce batch files for. Options are "windows" and "unix". Default is set to "windows".}

\item{foldername}{character name of the output folder with all MrBayes NEXUS files.}

\item{filename}{character, name of the output file.}
}
\value{
All necessary input files, organized in the GALAX/INPUT/foldername folder,
}
\description{
Builds all input files to run all analyses from resamp_bayes in Galax.
}
