% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp_info.R
\name{resamp_info}
\alias{resamp_info}
\title{Import and summarize results from analyses set by the resamp_galax function.}
\usage{
resamp_info(roots = roots, nsamp = 100, runs = 2, foldername = "resamp")
}
\arguments{
\item{roots}{character, a named vector with the ontology IDs of the anatomical data subsets to produce MrBayes NEXUS files. 
IDs and names should match those in the second and third columns of the ids data.frame, respectively.}

\item{nsamp}{integer, sets the number of samples for the resampling analysis.}

\item{runs}{integer, sets the number of independent runs of mcmc in the MrBayes commands block. Default is set to 2.}

\item{foldername}{character name of the output folder with all Galax output files.}
}
\value{
A list of matrices with values posterior coverage, Bayesian phylogenetic information (BPI) content, and phylogenetic dissonance
  among mcmc runs and among data subsets (if roots include more than one data subset) for each data subset specified in roots.
}
\description{
Import and summarize results from analyses set by the resamp_galax function.
}
